migrated remaining events, removed TorEventListener
