+++ orbotservice/src/main/java/org/torproject/android/service/util/ExternalIPFetcher.java
@@ -2,8 +2,8 @@ package org.torproject.android.service.util;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.torproject.android.service.OrbotRawEventListener;
 import org.torproject.android.service.OrbotService;
-import org.torproject.android.service.TorEventHandler;
 
 import java.io.BufferedReader;
 import java.io.InputStream;
@@ -17,10 +17,10 @@ public class ExternalIPFetcher implements Runnable {
 
     private final static String ONIONOO_BASE_URL = "https://onionoo.torproject.org/details?fields=country_name,as_name,or_addresses&lookup=";
     private OrbotService mService;
-    private TorEventHandler.Node mNode;
+    private OrbotRawEventListener.Node mNode;
     private int mLocalHttpProxyPort = 8118;
 
-    public ExternalIPFetcher(OrbotService service, TorEventHandler.Node node, int localProxyPort) {
+    public ExternalIPFetcher(OrbotService service, OrbotRawEventListener.Node node, int localProxyPort) {
         mService = service;
         mNode = node;
         mLocalHttpProxyPort = localProxyPort;
